+++ packages/zarm/src/marquee/interface.ts
@@ -2,7 +2,7 @@ export interface BaseMarqueeProps {
   direction: 'left' | 'right' | 'up' | 'down';
   loop: boolean;
   speed: number;
-  animationDelay: number;
+  delay: number;
   height?: number;
   width?: number;
 }