+++ packages/zarm/CHANGELOG.md
@@ -6,7 +6,7 @@
 
   - [Input] 拆分使用到虚拟键盘的自定义输入框组件 `CustomInput`。触发的事件入参由 value 值改为 event，受影响的事件有 `onChange` `onFocus` `onBlur` `onFocus`。删除 `onClear` 事件，点击重置按钮请空值将触发 `onChange` 事件。
   - [SearchBar] 删除 `onClear` 事件，点击重置按钮请空值将触发 `onChange` 事件。
-  - [Marquee] 用 `speed` 替换了 `animationDuration`， 解决因为内容长短导致速度不一致的问题。
+  - [Marquee] 用 `speed` 替换了 `animationDuration`， 解决因为内容长短导致速度不一致的问题。`delay` 代替了 `animationDelay`
 
 ## v2.9.2